Lint/refactor auth components
