+++ src/components/auth/ForgotPasswordForm.tsx
@@ -1,0 +1,105 @@
+import { useState } from 'react';
+
+import { Button } from '@/components/ui/button';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+import { supabase } from '@/lib/supabaseClient';
+import { cn } from '@/lib/utils';
+
+export interface ForgotPasswordFormProps extends React.ComponentPropsWithoutRef<'div'> {
+  onSuccess: () => void;
+  onViewChange: () => void;
+}
+
+/**
+ * Form to request a password reset email via Supabase Auth.
+ * Handles loading states, error reporting, and a success confirmation view.
+ *
+ * Side Effects:
+ * - Triggers a network request to Supabase to send a password reset email.
+ */
+export function ForgotPasswordForm({
+  className,
+  onSuccess,
+  onViewChange,
+  ...props
+}: React.ComponentPropsWithoutRef<'div'> & { onSuccess: () => void; onViewChange: () => void }) {
+  const [email, setEmail] = useState('');
+  const [error, setError] = useState<string | null>(null);
+  const [success, setSuccess] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
+
+  const handleForgotPassword = async (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    setIsLoading(true);
+    setError(null);
+    setSuccess(false);
+
+    try {
+      const { error } = await supabase.auth.resetPasswordForEmail(email);
+      if (error) throw error;
+      setSuccess(true);
+    } catch (error: unknown) {
+      setError(error instanceof Error ? error.message : 'An error occurred');
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  return (
+    <div className={cn('flex flex-col gap-6', className)} {...props}>
+      {success ? (
+        <Card>
+          <CardHeader>
+            <CardTitle className="text-2xl">Check Your Email</CardTitle>
+            <CardDescription>Password reset instructions sent.</CardDescription>
+          </CardHeader>
+          <CardContent className="flex flex-col gap-4">
+            <p className="text-sm text-muted-foreground">
+              If you registered using your email and password, you will receive a password reset email.
+            </p>
+            <Button className="w-full" onClick={onSuccess}>
+              Close
+            </Button>
+          </CardContent>
+        </Card>
+      ) : (
+        <Card>
+          <CardHeader>
+            <CardTitle className="text-2xl">Reset Your Password</CardTitle>
+            <CardDescription>Type in your email and we&apos;ll send you a link to reset your password</CardDescription>
+          </CardHeader>
+          <CardContent>
+            <form onSubmit={handleForgotPassword}>
+              <div className="flex flex-col gap-6">
+                <div className="grid gap-2">
+                  <Label htmlFor="email">Email</Label>
+                  <Input
+                    id="email"
+                    type="email"
+                    placeholder="m@example.com"
+                    required
+                    value={email}
+                    onChange={(e) => setEmail(e.target.value)}
+                    autoComplete="email"
+                  />
+                </div>
+                {error && <p className="text-sm text-red-500">{error}</p>}
+                <Button type="submit" className="w-full" disabled={isLoading}>
+                  {isLoading ? 'Sending...' : 'Send reset email'}
+                </Button>
+              </div>
+              <div className="mt-4 text-center text-sm">
+                Already have an account?{' '}
+                <Button type="button" variant="link" className="p-0 h-auto font-normal" onClick={onViewChange}>
+                  Login
+                </Button>
+              </div>
+            </form>
+          </CardContent>
+        </Card>
+      )}
+    </div>
+  );
+}
