+++ src/components/auth/UpdatePasswordForm.tsx
@@ -4,31 +4,40 @@ import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
-// import { createClient } from '@/lib/supabase/client'
 import { supabase } from '@/lib/supabaseClient';
 import { cn } from '@/lib/utils';
 
-export function ForgotPasswordForm({
+export interface UpdatePasswordFormProps extends React.ComponentPropsWithoutRef<'div'> {
+  onSuccess: () => void;
+}
+
+/**
+ * Renders a form for authenticated users to update their password via Supabase.
+ * Handles loading states, error reporting, and a success confirmation view.
+ *
+ * Side Effects:
+ * - Mutates the user's authentication credentials in Supabase.
+ */
+export function UpdatePasswordForm({
   className,
   onSuccess,
-  onViewChange,
   ...props
-}: React.ComponentPropsWithoutRef<'div'> & { onSuccess: () => void; onViewChange: () => void }) {
-  const [email, setEmail] = useState('');
+}: React.ComponentPropsWithoutRef<'div'> & { onSuccess: () => void }) {
+  const [password, setPassword] = useState('');
   const [error, setError] = useState<string | null>(null);
-  const [success, setSuccess] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
+  const [isSuccess, setIsSuccess] = useState(false);
 
-  const handleForgotPassword = async (e: React.FormEvent) => {
+  const handleUpdatePassword = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     setIsLoading(true);
     setError(null);
-    setSuccess(false);
+    setIsSuccess(false);
 
     try {
-      const { error } = await supabase.auth.resetPasswordForEmail(email);
+      const { error } = await supabase.auth.updateUser({ password });
       if (error) throw error;
-      setSuccess(true);
+      setIsSuccess(true);
     } catch (error: unknown) {
       setError(error instanceof Error ? error.message : 'An error occurred');
     } finally {
@@ -38,16 +47,13 @@ export function ForgotPasswordForm({
 
   return (
     <div className={cn('flex flex-col gap-6', className)} {...props}>
-      {success ? (
+      {isSuccess ? (
         <Card>
           <CardHeader>
-            <CardTitle className="text-2xl">Check Your Email</CardTitle>
-            <CardDescription>Password reset instructions sent.</CardDescription>
+            <CardTitle>Success</CardTitle>
+            <CardDescription>Your password has been updated successfully.</CardDescription>
           </CardHeader>
-          <CardContent className="flex flex-col gap-4">
-            <p className="text-sm text-muted-foreground">
-              If you registered using your email and password, you will receive a password reset email.
-            </p>
+          <CardContent>
             <Button className="w-full" onClick={onSuccess}>
               Close
             </Button>
@@ -57,32 +63,25 @@ export function ForgotPasswordForm({
         <Card>
           <CardHeader>
             <CardTitle className="text-2xl">Reset Your Password</CardTitle>
-            <CardDescription>Type in your email and we&apos;ll send you a link to reset your password</CardDescription>
+            <CardDescription>Please enter your new password below.</CardDescription>
           </CardHeader>
           <CardContent>
-            <form onSubmit={handleForgotPassword}>
+            <form onSubmit={handleUpdatePassword}>
               <div className="flex flex-col gap-6">
                 <div className="grid gap-2">
-                  <Label htmlFor="email">Email</Label>
+                  <Label htmlFor="password">New password</Label>
                   <Input
-                    id="email"
-                    type="email"
-                    placeholder="m@example.com"
+                    id="password"
+                    type="password"
+                    placeholder="New password"
                     required
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                    autoComplete="email"
+                    value={password}
+                    onChange={(e) => setPassword(e.target.value)}
                   />
                 </div>
                 {error && <p className="text-sm text-red-500">{error}</p>}
                 <Button type="submit" className="w-full" disabled={isLoading}>
-                  {isLoading ? 'Sending...' : 'Send reset email'}
-                </Button>
-              </div>
-              <div className="mt-4 text-center text-sm">
-                Already have an account?{' '}
-                <Button type="button" variant="link" className="p-0 h-auto font-normal" onClick={onViewChange}>
-                  Login
+                  {isLoading ? 'Saving...' : 'Save new password'}
                 </Button>
               </div>
             </form>
